Fixes for document creation: - work under Unix - correctly handle inheritance
git-svn-id: http://svn.code.sf.net/p/wxperl/code/dist-tools/trunk@1732 2e9104e8-d5c8-4c90-bb67-bfd1792b7337

+++ docs.pl
@@ -74,9 +74,11 @@ sub make_chm {
 
   my_chdir $html_build;
   die "no file 'manua.hhp'" unless -f 'manua.hhp';
-  system( "$hhc manua.hhp" );
-  die "no file 'manua.chm'" unless -f 'manua.chm';
-  my_system( "mv manua.chm $chmfile" );
+  if( -f $hhc ) {
+    system( "$hhc manua.hhp" );
+    die "no file 'manua.chm'" unless -f 'manua.chm';
+    my_system( "mv manua.chm $chmfile" );
+  }
 }
 
 sub make_tex_zip {
